Adding Source to newly created repository
Final Assignment:
Create a small e-commerce project that has 3 components:
- List of products
- Shopping Cart
- Order List

A person can add one or more products to a Shopping Cart.
A Shopping Cart displays products added to it and a total price for products in the Cart.
When user clicks a Check Out button in the Shopping Cart, the Order is created with the total order amount and is displayed in the Shopping Cart page.

Advanced: keep a list of bought products in the order.
All submitted orders are displayed below the Product List.

Products and Orders are saved and retrieved from the backend using REST services.